Move to separate file and add test
